Reduce repetition of id_dicts and use consistent setup names in e2e tests #413
