+++ test/e2e/test_catalogue_category.py
@@ -8,6 +8,7 @@ End-to-End tests for the catalogue category router.
 # pylint: disable=too-many-public-methods
 
 from test.e2e.conftest import E2ETestHelpers
+from test.e2e.test_unit import CreateDSL as UnitCreateDSL
 from test.mock_data import (
     CATALOGUE_CATEGORY_DATA_LEAF_NO_PARENT_WITH_PROPERTIES_MM,
     CATALOGUE_CATEGORY_GET_DATA_LEAF_NO_PARENT_NO_PROPERTIES,
@@ -31,47 +32,15 @@ from typing import Optional
 
 import pytest
 from bson import ObjectId
-from fastapi.testclient import TestClient
 from httpx import Response
 
 from inventory_management_system_api.core.consts import BREADCRUMBS_TRAIL_MAX_LENGTH
 
 
-class CreateDSL:
+class CreateDSL(UnitCreateDSL):
     """Base class for create tests."""
 
-    test_client: TestClient
-
     _post_response_catalogue_category: Response
-
-    unit_value_id_dict: dict[str, str]
-
-    @pytest.fixture(autouse=True)
-    def setup(self, test_client):
-        """Setup fixtures"""
-
-        self.test_client = test_client
-        self.unit_value_id_dict = {}
-
-    def add_unit_value_and_id(self, unit_value: str, unit_id: str) -> None:
-        """
-        Stores a unit value and ID inside the `unit_value_id_dict` for tests that need to have a
-        non-existent or invalid unit ID.
-
-        :param unit_value: Value of the unit.
-        :param unit_id: ID of the unit.
-        """
-
-        self.unit_value_id_dict[unit_value] = unit_id
-
-    def post_unit(self, unit_post_data: dict) -> None:
-        """Posts a unit with the given data and stores the value and ID in a dictionary for lookup later.
-
-        :param unit_post_data: Dictionary containing the unit data as would be required for a `UnitPostSchema`.
-        """
-
-        post_response = self.test_client.post("/v1/units", json=unit_post_data)
-        self.add_unit_value_and_id(unit_post_data["value"], post_response.json()["id"])
 
     def post_catalogue_category(self, catalogue_category_data: dict) -> Optional[str]:
         """
@@ -462,7 +431,7 @@ class GetBreadcrumbsDSL(GetDSL):
     _posted_catalogue_categories_get_data: list[dict]
 
     @pytest.fixture(autouse=True)
-    def setup_breadcrumbs_dsl(self):
+    def setup_catalogue_category_get_breadcrumbs_dsl(self):
         """Setup fixtures"""
 
         self._posted_catalogue_categories_get_data = []
